# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Abaya Production Tracking and Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cutting/batches:
    get:
      operationId: listCuttingBatches
      tags: [cutting]
      summary: List all cutting batches
      responses:
        "200":
          description: List of cutting batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CuttingBatch"
    post:
      operationId: createCuttingBatch
      tags: [cutting]
      summary: Create a cutting batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCuttingBatchBody"
      responses:
        "201":
          description: Created cutting batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CuttingBatch"
  /cutting/batches/{id}:
    get:
      operationId: getCuttingBatch
      tags: [cutting]
      summary: Get cutting batch details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cutting batch details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CuttingBatchDetail"
    put:
      operationId: updateCuttingBatch
      tags: [cutting]
      summary: Update cutting batch (admin only — safe fields)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCuttingBatchBody"
      responses:
        "200":
          description: Updated cutting batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CuttingBatch"

  # ALLOCATION
  /allocation:
    get:
      operationId: listAllocations
      tags: [allocation]
      summary: List all allocations
      responses:
        "200":
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allocation"
    post:
      operationId: createAllocation
      tags: [allocation]
      summary: Allocate cut pieces to stitcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationBody"
      responses:
        "201":
          description: Created allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
  /allocation/{id}:
    get:
      operationId: getAllocation
      tags: [allocation]
      summary: Get allocation details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Allocation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
    put:
      operationId: updateAllocation
      tags: [allocation]
      summary: Update allocation (admin only — safe fields)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAllocationBody"
      responses:
        "200":
          description: Updated allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"

  # RECEIVING
...
components:
  schemas:
...
    CuttingBatch:
      type: object
      properties:
        id:
          type: integer
        batchNumber:
          type: string
        productId:
          type: integer
        productCode:
          type: string
        productName:
          type: string
        fabricId:
          type: integer
        fabricCode:
          type: string
        fabricName:
          type: string
        materialId:
          type: integer
        materialCode:
          type: string
        materialName:
          type: string
        material2Id:
          type: integer
        material2Code:
          type: string
        material2Name:
          type: string
        sizeId:
          type: integer
        sizeName:
          type: string
        colorId:
          type: integer
        colorCode:
          type: string
        colorName:
          type: string
        itemCode:
          type: string
          description: Computed from product+fabric+material1+material2 codes joined with "-"
        quantityCut:
          type: integer
        availableForAllocation:
          type: integer
        totalAllocated:
          type: integer
        cutter:
          type: string
        cuttingDate:
          type: string
        remarks:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required: [id, batchNumber, quantityCut]
    CuttingBatchDetail:
      type: object
      properties:
        batch:
          $ref: "#/components/schemas/CuttingBatch"
        fabricRolls:
          type: array
          items:
            $ref: "#/components/schemas/BatchFabricUsage"
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/Allocation"
      required: [batch, fabricRolls, allocations]
    BatchFabricUsage:
      type: object
      properties:
        rollNumber:
          type: string
        fabricName:
          type: string
        quantityUsed:
          type: number
        unit:
          type: string
      required: [rollNumber, quantityUsed]
    CreateCuttingBatchBody:
      type: object
      properties:
        batchNumber:
          type: string
          description: Manually entered batch number, must be unique
        productId:
          type: integer
        fabricId:
          type: integer
        materialId:
          type: integer
        material2Id:
          type: integer
        sizeId:
          type: integer
        colorId:
          type: integer
        quantityCut:
          type: integer
        cutter:
          type: string
        cuttingDate:
          type: string
        remarks:
          type: string
        fabricUsages:
          type: array
          items:
            type: object
            properties:
              fabricRollId:
                type: integer
              quantityUsed:
                type: number
            required: [fabricRollId, quantityUsed]
      required: [batchNumber, productId, quantityCut, cuttingDate]
    UpdateCuttingBatchBody:
      type: object
      properties:
        cutter:
          type: string
        cuttingDate:
          type: string
        remarks:
          type: string

    # ALLOCATION
    Allocation:
      type: object
      properties:
        id:
          type: integer
        allocationNumber:
          type: string
        cuttingBatchId:
          type: integer
        batchNumber:
          type: string
        productCode:
          type: string
        productName:
          type: string
        fabricCode:
          type: string
        fabricName:
          type: string
        materialCode:
          type: string
        materialName:
          type: string
        material2Code:
          type: string
        material2Name:
          type: string
        itemCode:
          type: string
        sizeName:
          type: string
        colorCode:
          type: string
        colorName:
          type: string
        stitcherId:
          type: integer
        stitcherName:
          type: string
        teamName:
          type: string
        quantityIssued:
          type: integer
        quantityReceived:
          type: integer
        quantityPending:
          type: integer
        quantityRejected:
          type: integer
        issueDate:
          type: string
        remarks:
          type: string
        status:
          type: string
      required: [id, quantityIssued]
    CreateAllocationBody:
      type: object
      properties:
        cuttingBatchId:
          type: integer
        stitcherId:
          type: integer
        quantityIssued:
          type: integer
        issueDate:
          type: string
        remarks:
          type: string
      required: [cuttingBatchId, stitcherId, quantityIssued, issueDate]
    UpdateAllocationBody:
      type: object
      properties:
        issueDate:
          type: string
        remarks:
          type: string

    # RECEIVING
...