# lib/api-zod/src/generated/types/allocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Abaya Production Tracking and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Allocation {
  id: number;
  allocationNumber?: string;
  cuttingBatchId?: number;
  batchNumber?: string;
  productCode?: string;
  productName?: string;
  fabricCode?: string;
  fabricName?: string;
  materialCode?: string;
  materialName?: string;
  material2Code?: string;
  material2Name?: string;
  itemCode?: string;
  sizeName?: string;
  colorCode?: string;
  colorName?: string;
  stitcherId?: number;
  stitcherName?: string;
  teamName?: string;
  quantityIssued: number;
  quantityReceived?: number;
  quantityPending?: number;
  quantityRejected?: number;
  issueDate?: string;
  remarks?: string;
  status?: string;
}