# lib/api-zod/src/generated/types/createCuttingBatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Abaya Production Tracking and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCuttingBatchBodyFabricUsagesItem } from "./createCuttingBatchBodyFabricUsagesItem";

export interface CreateCuttingBatchBody {
  /** Manually entered batch number, must be unique */
  batchNumber: string;
  productId: number;
  fabricId?: number;
  materialId?: number;
  material2Id?: number;
  sizeId?: number;
  colorId?: number;
  quantityCut: number;
  cutter?: string;
  cuttingDate: string;
  remarks?: string;
  fabricUsages?: CreateCuttingBatchBodyFabricUsagesItem[];
}