# lib/api-zod/src/generated/types/cuttingBatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Abaya Production Tracking and Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CuttingBatch {
  id: number;
  batchNumber: string;
  productId?: number;
  productCode?: string;
  productName?: string;
  fabricId?: number;
  fabricCode?: string;
  fabricName?: string;
  materialId?: number;
  materialCode?: string;
  materialName?: string;
  material2Id?: number;
  material2Code?: string;
  material2Name?: string;
  sizeId?: number;
  sizeName?: string;
  colorId?: number;
  colorCode?: string;
  colorName?: string;
  /** Computed from product+fabric+material1+material2 codes joined with "-" */
  itemCode?: string;
  quantityCut: number;
  availableForAllocation?: number;
  totalAllocated?: number;
  cutter?: string;
  cuttingDate?: string;
  remarks?: string;
  status?: string;
  createdAt?: string;
}